Removing nexus_wrapper.py entirely
