+++ nexus_constructor/kafka/status_consumer.py
@@ -3,9 +3,7 @@ import logging
 import time
 from copy import copy
 from uuid import uuid1
-
 import confluent_kafka
-
 from nexus_constructor.kafka.kafka_interface import KafkaInterface
